Rename zmq to zeromq module
